Extract changeStatus to its own file.
